Use custom host whenever its defined
Links, tags, imgs, etc. should use the defined host whenver its defined, regardless of whether or not the server is running in production mode or development mode. If its desired to set a host in production, but not development, that should be set in configuration files.

Also added a link tag helper to create links with the appended hosts for non-stylesheet link tags